fix h4x node maps (again?)

+++ Adjutant/Blam/Halo4/render_model.cs
@@ -202,7 +202,7 @@ namespace Adjutant.Blam.Halo4
     }
 
     [Flags]
-    public enum ModelFlags : int
+    public enum ModelFlags : short
     {
         UseLocalNodes = 4
     }
